requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyJWT==2.8.0
pytest==7.4.3
pytest-flask==1.3.0